Special treatment for extraLineFragment for empty document
TextLayoutManager layout extra line fragment for empty document, however, it has no information about text attribute itself. This result in layout based on a default text style, that is different than text view default (typing) set of attributes. All of this result in measurements of the layout fragments that are unexpected in the context