Change undo behaviour (#53)
This pull request changes the behaviour of the undo functionality and
fixes some small issues in the current implementation:
* If you edited at the end of the current text (e.g. by typing at the
end of the text) and used undo, then the undo didn't work properly and
only the cursor moved, but the text was not removed again. This happened
because the undo range was calculated before the text was added,
therefore the undo range was always too small.
* If you pasted text that was longer than the end of the document and
used undo, then the text ended up in a wrong state and parts of the text
were repeated.
* An undo operation could happen at the wrong place in the text.

Unlike before, a newline will now break the typing coalescing which is
the same behaviour as in other text editors (e.g. Xcode or TextEdit).

Also, an undo operation will now restore the text selection that existed
when the operation that is undone was done, again mimicking the
behaviour of other editors such as Xcode.

Additionally, this pull request provides some handy helpers to create
keyboard events that can be used in tests.

---------

Co-authored-by: Lukas Stührk <lu.stuehrk@mabah02jh4h6nk.fritz.box>