Preserves position in line fragments when moving up and down (#57)
Ensures the location is preserved when navigating up and down between
line fragments using the keyboard. The behavior matches UITextView. The
videos below compares the behavior before and after these changes.

The implementation uses
[anchorPositionOffset](https://developer.apple.com/documentation/uikit/nstextselection/3801817-anchorpositionoffset?changes=la_3)
on NSTextSelection. TextKit 2 automatically populates this on macOS but
we need to do it manually on iOS to achieve the same behavior.