Add gutter markers
Add new STGutterMarker type to represent a marker displayed in the gutter.
Markers can be added, removed, and positioned next to the corresponding line
number cell. Tapping the gutter toggles a default marker for that line.

The markers are displayed in a new dedicated view on top of the line number
cells. Marker layout is updated when markers are added or removed.