Add licensing information and commercial license alternative
This change adds licensing information to the top of the LICENSE.md file,
clarifying that the software is copyrighted by Marcin Krzyżanowski and
governed by the included license. It also mentions that commercial licenses
are available as an alternative to the GPL v3.0 for organizations that
require it.

The added text provides context on the software's licensing and points to
the commercial license option for cases where the open source license is
not suitable. This helps inform users of their licensing options and
obligations when using the STTextView software.