Accept dragged text from within and outside the text view (#89)
This implements accepting dragged text from within the text view and
from other apps.
It follows from discussion #87.

Dragging in from other apps is quite straightforward, it uses the
existing `updateTextSelection` and `readSelection` functions to show the
drop location and insert any text.

Dragging within the view is more manual as we've got to gather the text,
figure out how the insertion point changes after removing that text,
then perform the removal and insertion.

It should be able to cope with dragging from multiple selections, but
the view doesn't seem to start a dragging session when there are
multiple separate selections. I'm not sure where to change that, but
don't think it's required to allow this change as a starting point.

I'm not very familiar with the TextKit 2 APIs, so my implementation of
`performInternalDragOperation` could probably be improved a fair bit.
Mostly around the conversions between `NSTextRange` and `NSRange`. I'd
welcome any suggestions for improvement.